requirements.txt
# Phase 1: Data Pull & Freezing
requests>=2.28.0

# Phase 2: Feature Extraction
torch>=2.0.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb
Pillow>=10.3.0
numpy>=1.24.0

# Phase 2B: Caption Embeddings
sentence-transformers>=2.2.0

# Phase 2C: Emotion Extraction
transformers>=4.30.0,<5.0.0

# Phase 2E: Location Clustering
scikit-learn>=1.3.0

# Vector Database
chromadb>=0.4.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocs-minify-plugin>=0.7.0

# Grand Fusion (optional: only needed for --export to Parquet)
pandas>=2.0.0
pyarrow>=14.0.1

# Location Semantic
httpx==0.28.1
google-genai==1.5.0
python-dotenv==1.0.1
pydantic==2.10.6

